Consolidate common code
