+++ GreekReference/src/main/java/com/benlinskey/greekreference/AbstractDetailActivity.java
@@ -78,28 +78,12 @@ public abstract class AbstractDetailActivity extends ContainerActivity {
      */
     protected abstract void restoreActionBar();
 
-    // The following two methods are a workaround for a bug related to the appcompat-v7 library
-    // on some LG devices. Thanks to Alex Lockwood for the fix: 
-    // http://stackoverflow.com/questions/26833242/nullpointerexception-phonewindowonkeyuppanel1002-main
-
-    @Override
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
-            return true;
-        }
-        return super.onKeyDown(keyCode, event);
-    }
-
-    @Override
-    public boolean onKeyUp(int keyCode, KeyEvent event) {
-        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
-            openOptionsMenu();
-            return true;
-        }
-        return super.onKeyUp(keyCode, event);
-    }
-
     protected AbstractDetailFragment getDetailFragment() {
         return (AbstractDetailFragment) getFragmentManager().findFragmentById(R.id.item_detail_container);
+    }
+
+    public void displayDetailViewToast(String msg) {
+        AbstractDetailFragment fragment = getDetailFragment();
+        fragment.displayToast(msg);
     }
 }
