+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/ContainerActivity.java
@@ -22,10 +22,12 @@ import android.app.DialogFragment;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 import android.text.Html;
 import android.text.method.LinkMovementMethod;
+import android.view.KeyEvent;
 import android.widget.ScrollView;
 import android.widget.TextView;
 
@@ -75,6 +77,27 @@ public class ContainerActivity extends ActionBarActivity {
         Intent intent = new Intent(Intent.ACTION_SENDTO, uri);
         intent.putExtra(Intent.EXTRA_SUBJECT, getString(R.string.feedback_subject));
         startActivity(Intent.createChooser(intent, getString(R.string.feedback_intent_chooser)));
+    }
+
+    // The following two methods are a workaround for a bug related to the appcompat-v7 library
+    // on some LG devices. Thanks to Alex Lockwood for the fix:
+    // http://stackoverflow.com/questions/26833242/nullpointerexception-phonewindowonkeyuppanel1002-main
+
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
+            return true;
+        }
+        return super.onKeyDown(keyCode, event);
+    }
+
+    @Override
+    public boolean onKeyUp(int keyCode, KeyEvent event) {
+        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
+            openOptionsMenu();
+            return true;
+        }
+        return super.onKeyUp(keyCode, event);
     }
 
 }
