+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/lexicon/LexiconDetailActivity.java
@@ -17,11 +17,9 @@
 package com.benlinskey.greekreference.views.lexicon;
 
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.NavUtils;
 import android.support.v7.app.ActionBar;
-import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuItem;
 
@@ -97,12 +95,6 @@ public class LexiconDetailActivity extends AbstractDetailActivity implements Lex
     }
 
     @Override
-    public void displayDetailViewToast(String msg) {
-        LexiconDetailFragment fragment = (LexiconDetailFragment) getDetailFragment();
-        fragment.displayToast(msg);
-    }
-
-    @Override
     public boolean isDetailFragmentVisible() {
         return true;
     }
@@ -132,23 +124,4 @@ public class LexiconDetailActivity extends AbstractDetailActivity implements Lex
         actionBar.setTitle(mTitle);
     }
 
-    // The following two methods are a workaround for a bug related to the appcompat-v7 library
-    // on some LG devices. Thanks to Alex Lockwood for the fix: http://stackoverflow.com/questions/26833242/nullpointerexception-phonewindowonkeyuppanel1002-main
-
-    @Override
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
-            return true;
-        }
-        return super.onKeyDown(keyCode, event);
-    }
-
-    @Override
-    public boolean onKeyUp(int keyCode, KeyEvent event) {
-        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
-            openOptionsMenu();
-            return true;
-        }
-        return super.onKeyUp(keyCode, event);
-    }
 }
