+++ GreekReference/src/main/java/com/benlinskey/greekreference/views/syntax/SyntaxDetailActivity.java
@@ -17,11 +17,9 @@
 package com.benlinskey.greekreference.views.syntax;
 
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.NavUtils;
 import android.support.v7.app.ActionBar;
-import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuItem;
 
@@ -117,31 +115,5 @@ public class SyntaxDetailActivity extends AbstractDetailActivity implements Synt
         actionBar.setTitle(mTitle);
     }
 
-    @Override
-    public void displayDetailViewToast(String msg) {
-        SyntaxDetailFragment fragment = (SyntaxDetailFragment) getDetailFragment();
-        fragment.displayToast(msg);
-    }
-
-    // The following two methods are a workaround for a bug related to the appcompat-v7 library
-    // on some LG devices. Thanks to Alex Lockwood for the fix: 
-    // http://stackoverflow.com/questions/26833242/nullpointerexception-phonewindowonkeyuppanel1002-main
-
-    @Override
-    public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
-            return true;
-        }
-        return super.onKeyDown(keyCode, event);
-    }
-
-    @Override
-    public boolean onKeyUp(int keyCode, KeyEvent event) {
-        if (KeyEvent.KEYCODE_MENU == keyCode && Build.BRAND.equalsIgnoreCase("LGE")) {
-            openOptionsMenu();
-            return true;
-        }
-        return super.onKeyUp(keyCode, event);
-    }
 }
